// apps/dispatcher-web/src/app/models/orders/history-orders.model.ts
export interface HistoryOrder {
    date: string;
    orderNo: string;
    customerName: string;
    pickup: string;
    amount: string;
    distance: string;
    placementTime: string;
    startTime: string;
    pickupTime: string;
    deliveryTime: string;
    driver: string;
    status: string;
}